+++ src/components/kit/flavorsModal/card/card.css
