+++ src/components/kit/flavorsModal/card/card.tsx
