+++ src/components/kit/flavorsModal/card/cardflavor.tsx
