+++ src/components/kit/flavorsModal/card/index.tsx
