+++ src/components/kit/flavorsModal/index.style.tsx
@@ -1,0 +1,22 @@
+export const flavorsPanel = {
+    display: 'flex',
+    flexWrap: 'wrap',
+    justifyContent: 'center',
+    overflowY: 'scroll',
+    height: '650px',
+    padding: '50px'
+}
+
+export const modalStyle = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: '90%',
+    bgcolor: 'background.paper',
+    boxShadow: 24,
+    display: 'flex',
+    flexWrap: 'wrap',
+    flexDirection: 'column',
+    borderRadius: '10px'
+};
