+++ src/components/kit/flavorsModal/index.types.tsx
@@ -1,0 +1,7 @@
+export interface IFlavor {
+    ativo: 0|1,
+    description: string,
+    id: number,
+    link: string,
+    nome: string
+}
