--- src/components/modal/index.tsx
@@ -1,3 +1,0 @@
-import FlavorsModal from "./modal";
-
-export default FlavorsModal
