+++ src/components/navbar/account/index.style.tsx
@@ -1,0 +1,16 @@
+export const boxModal = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    bgcolor: 'background.paper',
+    display: 'flex',
+    flexWrap: 'wrap',
+    flexDirection: 'column',
+    borderRadius: '10px'
+};
+  
+export const accountTabs = {
+    p: '12px 80px'
+}
+
