+++ src/components/navbar/account/index.types.tsx
@@ -1,0 +1,6 @@
+export interface TabPanelProps {
+    children?: React.ReactNode;
+    dir?: string;
+    index: number;
+    value: number;
+}
