+++ src/components/navbar/account/sign-up/Form/Address/CEP/index.tsx
@@ -1,0 +1,16 @@
+import { TextField } from "@mui/material";
+
+export function SignUpCEP(){
+
+    return (
+        <TextField
+        required
+        fullWidth
+        type="number"
+        id="cep"
+        label="CEP"
+        name="cep"
+        autoComplete="cep"
+      />
+    )
+}
