+++ src/components/navbar/account/sign-up/Form/Address/Complement/index.tsx
@@ -1,0 +1,13 @@
+import { TextField } from "@mui/material";
+
+export function SignUpAddressComplement() {
+    return (
+        <TextField
+            fullWidth
+            id="complemento"
+            label="Complemento"
+            name="complemento"
+            autoComplete="complemento"
+        />
+    )
+}
