+++ src/components/navbar/account/sign-up/Form/Address/District/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignUpAddressDistrict() {
+    return (
+        <TextField
+            required
+            fullWidth
+            id="bairro"
+            label="Bairro"
+            name="bairro"
+            type="text"
+            autoComplete="bairro"
+        />
+    )
+}
