+++ src/components/navbar/account/sign-up/Form/Address/Number/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignUpAddressNumber() {
+    return (
+        <TextField
+            required
+            fullWidth
+            type="number"
+            id="numero"
+            label="Número"
+            name="numero"
+            autoComplete="numero"
+        />
+    )    
+}
