+++ src/components/navbar/account/sign-up/Form/Address/PublicPlace/index.tsx
@@ -1,0 +1,14 @@
+import { TextField } from "@mui/material";
+
+export function SignUpAddressPublicPlace() {
+    return (
+        <TextField
+            required
+            fullWidth
+            id="logradouro"
+            label="Logradouro"
+            name="logradouro"
+            autoComplete="logradouro"
+        />
+    )
+}
