+++ src/components/navbar/account/sign-up/Form/ConfirmPassword/Index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignUpConfirmPassword() {
+    return (
+        <TextField
+            required
+            fullWidth
+            name="confirm_password"
+            label="Confirm Password"
+            type="password"
+            id="confirm_password"
+            autoComplete="new-confirm_password"
+        />
+    )
+}
