+++ src/components/navbar/account/sign-up/Form/Email/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignupEmail(){
+    return (
+        <TextField
+            required
+            fullWidth
+            id="email"
+            label="E-mail"
+            name="email"
+            type="email"
+            autoComplete="email"
+        />
+    )
+}
