+++ src/components/navbar/account/sign-up/Form/Name/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignUpName(){
+    return (
+        <TextField
+            autoComplete="given-name"
+            name="completeName"
+            required
+            fullWidth
+            id="completeName"
+            label="Nome Completo"
+            autoFocus
+        />
+    );
+}
