+++ src/components/navbar/account/sign-up/Form/Password/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignUpPassword() {
+    return (
+        <TextField
+            required
+            fullWidth
+            name="password"
+            label="Password"
+            type="password"
+            id="password"
+            autoComplete="new-password"
+        />
+    )
+}
