+++ src/components/navbar/account/sign-up/Form/Phone/index.tsx
@@ -1,0 +1,15 @@
+import { TextField } from "@mui/material";
+
+export function SignupPhone() {
+    return (
+        <TextField
+            required
+            fullWidth
+            id="phone"
+            label="Celular"
+            name="phone"
+            type="number"
+            autoComplete="phone"
+        />
+    )
+}
