--- src/components/navbar/navbar.css
@@ -1,7 +1,0 @@
-.navbar {
-     
-}
-
-.accountTabs {
-    padding: 12px 30px;
-}
